add printing lock, huge pages keep track of size

+++ src/med_malloc.c
@@ -6,7 +6,7 @@
 /*   By: kyork <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/22 18:18:15 by kyork             #+#    #+#             */
-/*   Updated: 2018/05/29 17:49:51 by kyork            ###   ########.fr       */
+/*   Updated: 2018/05/29 18:55:49 by kyork            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,14 +62,14 @@ static ssize_t			bitmask_claim(ATOM_U64 *set, t_u8 size, int max)
 
 static void				*reservem(t_region *page, int size)
 {
-	ssize_t			idx;
+	size_t			idx;
 	ssize_t			r;
 
 	idx = 0;
 	while (idx < page->item_count)
 	{
 		r = bitmask_claim(pg_bitset_ptr(page, idx), size,
-				(page->item_count - idx));
+				(((ssize_t)page->item_count) - idx));
 		if (r != -1)
 			break ;
 		idx += 32;
